replaceComponents on splitEdge too
